do not use HMR in prod

+++ src/bot.ts
@@ -69,9 +69,10 @@ client.registry
 client.login(process.env.BOT_TOKEN);
 
 declare const module: any;
-// You need only 3 lines of code to start accepting code changes coming through the HMR
-if (module.hot) {
-  module.hot.accept();
-  // Next callback is essential: After code changes were accepted     we need to restart the app. server.close() is here Express.JS-specific and can differ in other frameworks. The idea is that you should shut down your app here. Data/state saving between shutdown and new start is possible
-  module.hot.dispose(() => client.destroy());
+// Do not HMR in production
+if (process.env.NODE_ENV !== 'production') {
+  if (module.hot) {
+    module.hot.accept();
+    module.hot.dispose(() => client.destroy());
+  }
 }
